Adds bracket push soluction
Implementing this recursive solution—at least in
this way—meant defining default arguments and
handling a couple of base cases.  This meant
defining a function head and defining the method
again for each of the base cases.  Plus the default
implementation.  This doesn't seem overly readable,
as one needs to look in four places to really
understand how the method works.